Добавила контроллер для отрисовки точек маршрута
